+++ lib/widgets/fields/form_chips.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:nylo_support/localization/app_localization.dart';
 import '/helpers/extensions.dart';
 import '/widgets/fields/field_base_state.dart';
 import '/widgets/ny_form.dart';
@@ -86,7 +87,7 @@ class _NyFormChipState extends FieldBaseState<NyFormChip> {
                       isSelected ? getSelectedSide() : getUnselectedSide(),
                   dark: () => BorderSide(color: Colors.transparent)),
               shape: getShape(),
-              label: Text(option,
+              label: Text(option.tr(),
                   style: whenTheme(
                       light: () => isSelected
                           ? getSelectedTextStyle()